lge_battery: Defer probe if battery ID is zero
A battery ID is only going to be zero if the resistance is
returned as zero by fg_get_battery_resistance, which appears
to only happen when the battery ID hasn't been loaded, according
to logging on boot.

This avoids the -EINVAL return trigger below since commit
0277be04135e ("qpnp-fg-gen3: Move getting battery id and
profile to profile_load_work") delays the battery ID getting
loaded.

Change-Id: I204e6922598512572897666236467654fc04248a
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>